Tree and Lock classes isolated
