--- src/io.cpp
@@ -1,5 +1,0 @@
-#include "chess.h"
-
-Lock IOLock;
-
-
