+++ src/timer.h
@@ -1,3 +1,6 @@
+#ifndef timer_h__
+#define timer_h__
+
 #include <time.h>
 
 typedef unsigned long int TimerType;
@@ -13,4 +16,4 @@ public:
   TimerType ElapsedTime();
 };
 
-
+#endif  //timer_h__
