+++ src/tree/print_tree.cpp
@@ -1,4 +1,4 @@
-#include "chess.h"
+#include "tree.h"
 
 
 void ResetTree::PrintChildren()
